requirements.txt
py-cord==2.3.2
python-dotenv==0.21.0
dnspython==2.2.1
